code/python/appium-applet-ui/requirements.txt
adbutils==0.8.2
alembic==1.4.3
allure-pytest==2.8.24
allure-python-commons==2.8.24
apkutils2==1.0.0
Appium-Python-Client==1.0.2
attrdict==2.0.1
attrs==19.3.0
cached-property==1.5.2
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
cigam==0.0.3
click==7.1.2
ddt==1.4.2
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
facebook-wda==1.2.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
html-testRunner==1.2.1
idna==2.10
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
logzero==1.6.2
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pi==0.1.2
Pillow==8.0.1
pluggy==0.13.1
progress==1.5
py==1.8.1
pyelftools==0.27
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
retry==0.9.2
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.1
uiautomator2==2.12.1.dev1
unittestreport==1.1.10
urllib3==1.25.9
wcwidth==0.2.4
weditor==0.6.1
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
zipp==3.1.0
